requirements.txt
pypdf==5.1.0
google-generativeai
python-dotenv
streamlit
